+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.32.0] - 2024-08-06
+
 ### Added
 
 - Add experimental "Installations" page, visible with feature flag `show-installations-page`.
@@ -582,7 +584,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Disable anonymous access.
 
-[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.31.0...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.32.0...HEAD
+[0.32.0]: https://github.com/giantswarm/backstage/compare/v0.31.0...v0.32.0
 [0.31.0]: https://github.com/giantswarm/backstage/compare/v0.30.0...v0.31.0
 [0.30.0]: https://github.com/giantswarm/backstage/compare/v0.29.0...v0.30.0
 [0.29.0]: https://github.com/giantswarm/backstage/compare/v0.28.0...v0.29.0